convert test_class_mapping to use persistence
